Avoid OtterAPI's searchcount feature.
Before: We were using the searchcount feature, which was returning
results inconsistent with the actual search. That led to an ugly
workaround when we were processing the actual search results.

After: We use the search feature both for actually pulling a tip *and*
for calculating a page and index for said tip.