Minor cleanups.
Before: I was importing some modules too inside a method, but it's
pretty much a guarantee I need it for every request; this was
suboptimal. There was also a variable that was more like an invariant.

After: I moved all three to the top of the file, and then I improved the
variable's name while I was at it.